# specs/001-docusaurus-docs-site/contracts/openapi-backend.yaml
openapi: 3.0.0
info:
  title: AI Textbook Backend API
  version: 1.0.0
  description: Backend API for AI-native textbook with RAG chatbot, auth, translation
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api-textbook.example.com
    description: Production
paths:
...
  /chat/ask:
    post:
      summary: Ask chatbot a question
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                question: { type: string }
      responses:
        '200':
          description: Answer generated
...